requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# D:\Work\Python\Bagmati_TOF_App\pages\.ipynb_checkpoints\District_data-checkpoint.py: 8
# D:\Work\Python\Bagmati_TOF_App\pages\District_data.py: 8
Pillow == 9.0.1

# D:\Work\Python\Bagmati_TOF_App\.ipynb_checkpoints\folium_map2-checkpoint.py: 14
# D:\Work\Python\Bagmati_TOF_App\.ipynb_checkpoints\main-checkpoint.py: 7
# D:\Work\Python\Bagmati_TOF_App\main.py: 7
folium == 0.12.1.post1

# D:\Work\Python\Bagmati_TOF_App\.ipynb_checkpoints\folium_map2-checkpoint.py: 6
# D:\Work\Python\Bagmati_TOF_App\.ipynb_checkpoints\main-checkpoint.py: 3
# D:\Work\Python\Bagmati_TOF_App\main.py: 3
geopandas == 0.10.2

# D:\Work\Python\Bagmati_TOF_App\.ipynb_checkpoints\folium_map2-checkpoint.py: 12
matplotlib == 3.5.1

# D:\Work\Python\Bagmati_TOF_App\.ipynb_checkpoints\folium_map2-checkpoint.py: 3
numpy == 1.22.4

# D:\Work\Python\Bagmati_TOF_App\pages\.ipynb_checkpoints\District_data-checkpoint.py: 5
# D:\Work\Python\Bagmati_TOF_App\pages\District_data.py: 5
opencv-contrib-python == 4.5.5.64

# D:\Work\Python\Bagmati_TOF_App\pages\.ipynb_checkpoints\District_data-checkpoint.py: 5
# D:\Work\Python\Bagmati_TOF_App\pages\District_data.py: 5
opencv-contrib-python-headless == 4.5.5.64

# D:\Work\Python\Bagmati_TOF_App\pages\.ipynb_checkpoints\District_data-checkpoint.py: 5
# D:\Work\Python\Bagmati_TOF_App\pages\District_data.py: 5
opencv-openvino-contrib-python == 4.1.1.26

# D:\Work\Python\Bagmati_TOF_App\pages\.ipynb_checkpoints\District_data-checkpoint.py: 5
# D:\Work\Python\Bagmati_TOF_App\pages\District_data.py: 5
opencv-python == 4.5.5.64

# D:\Work\Python\Bagmati_TOF_App\pages\.ipynb_checkpoints\District_data-checkpoint.py: 5
# D:\Work\Python\Bagmati_TOF_App\pages\District_data.py: 5
opencv-python-aarch64 == 3.3.0.1

# D:\Work\Python\Bagmati_TOF_App\pages\.ipynb_checkpoints\District_data-checkpoint.py: 5
# D:\Work\Python\Bagmati_TOF_App\pages\District_data.py: 5
opencv-python-headless == 4.5.5.64

# D:\Work\Python\Bagmati_TOF_App\pages\.ipynb_checkpoints\District_data-checkpoint.py: 5
# D:\Work\Python\Bagmati_TOF_App\pages\District_data.py: 5
opencv-python-inference-engine == 2022.1.5

# D:\Work\Python\Bagmati_TOF_App\pages\.ipynb_checkpoints\District_data-checkpoint.py: 5
# D:\Work\Python\Bagmati_TOF_App\pages\District_data.py: 5
opencvz-contrib-python == 4.1.1.26

# D:\Work\Python\Bagmati_TOF_App\.ipynb_checkpoints\folium_map2-checkpoint.py: 2
# D:\Work\Python\Bagmati_TOF_App\.ipynb_checkpoints\main-checkpoint.py: 2
# D:\Work\Python\Bagmati_TOF_App\main.py: 2
pandas == 1.4.2

# D:\Work\Python\Bagmati_TOF_App\.ipynb_checkpoints\folium_map2-checkpoint.py: 10
plotly == 5.8.0

# D:\Work\Python\Bagmati_TOF_App\.ipynb_checkpoints\folium_map2-checkpoint.py: 4
pydeck == 0.7.1

# D:\Work\Python\Bagmati_TOF_App\.ipynb_checkpoints\folium_map2-checkpoint.py: 8,9
# D:\Work\Python\Bagmati_TOF_App\.ipynb_checkpoints\main-checkpoint.py: 5
# D:\Work\Python\Bagmati_TOF_App\main.py: 5
pyproj == 3.3.1

# D:\Work\Python\Bagmati_TOF_App\.ipynb_checkpoints\folium_map2-checkpoint.py: 1
# D:\Work\Python\Bagmati_TOF_App\.ipynb_checkpoints\main-checkpoint.py: 1
# D:\Work\Python\Bagmati_TOF_App\main.py: 1
# D:\Work\Python\Bagmati_TOF_App\pages\.ipynb_checkpoints\District_data-checkpoint.py: 1
# D:\Work\Python\Bagmati_TOF_App\pages\District_data.py: 1
streamlit == 1.10.0

# D:\Work\Python\Bagmati_TOF_App\.ipynb_checkpoints\folium_map2-checkpoint.py: 13
# D:\Work\Python\Bagmati_TOF_App\.ipynb_checkpoints\main-checkpoint.py: 6
# D:\Work\Python\Bagmati_TOF_App\main.py: 6
streamlit_folium == 0.6.12
